+++ hooks/useOpenAI.ts
@@ -7,6 +7,8 @@ import { create } from "zustand";
 type State = {
   openAIKey: string;
   envOpenAIKey: string;
+  azureOpenAIKey: string;
+
   /**
    * What sampling temperature to use, between 0 and 2.
    * Higher values like 0.8 will make the output more random,
@@ -42,21 +44,38 @@ export type UseOpenAIReturn = [State, (args: SaveOpenAI) => void];
 
 const modelOptions = [
   {
-    label: "gpt-3.5-turbo-0301",
-    value: "gpt-3.5-turbo-0301",
+    label: "OpenAI",
+    value: "openai",
+    children: [
+      {
+        label: "gpt-3.5-turbo-0301",
+        value: "openai-gpt-3.5-turbo-0301",
+      },
+      {
+        label: "gpt-3.5-turbo",
+        value: "openai-gpt-3.5-turbo",
+      },
+    ],
   },
   {
-    label: "gpt-3.5-turbo",
-    value: "gpt-3.5-turbo",
+    label: "Azure OpenAI",
+    value: "azure-openai",
+    children: [
+      {
+        label: "gpt-3.5-turbo-0301",
+        value: "azure-gpt-3.5-turbo-0301",
+      },
+    ],
   },
 ];
 
 const useStore = create<State & Action>((set) => ({
   openAIKey: "",
   envOpenAIKey: "",
+  azureOpenAIKey: "",
   temperature: 1,
   max_tokens: 2000,
-  model: "gpt-3.5-turbo",
+  model: "openai-gpt-3.5-turbo",
 
   update: (args: SaveOpenAI) => {
     if (typeof args === "function") {
@@ -64,25 +83,36 @@ const useStore = create<State & Action>((set) => ({
         const newState = JSON.parse(
           JSON.stringify({
             openAIKey: state.openAIKey,
+            azureOpenAIKey: state.azureOpenAIKey,
             model: state.model,
             temperature: state.temperature,
             max_tokens: state.max_tokens,
           })
         );
-        const { openAIKey, model, temperature, max_tokens } = args(newState);
+        const { openAIKey, azureOpenAIKey, model, temperature, max_tokens } =
+          args(newState);
         localStorage.setItem("openaiKey", openAIKey);
+        localStorage.setItem("azureOpenAIKey", azureOpenAIKey);
         localStorage.setItem("language_model", model);
         localStorage.setItem("temperature", String(temperature));
         localStorage.setItem("max_tokens", String(max_tokens));
-        return { openAIKey, model, temperature, max_tokens };
+        return { openAIKey, azureOpenAIKey, model, temperature, max_tokens };
       });
     } else {
-      const { openAIKey, model, temperature, max_tokens } = args;
+      const { openAIKey, azureOpenAIKey, model, temperature, max_tokens } =
+        args;
       localStorage.setItem("openaiKey", openAIKey);
+      localStorage.setItem("azureOpenAIKey", azureOpenAIKey);
       localStorage.setItem("language_model", model);
       localStorage.setItem("temperature", String(temperature));
       localStorage.setItem("max_tokens", String(max_tokens));
-      set(() => ({ openAIKey, model, temperature, max_tokens }));
+      set(() => ({
+        openAIKey,
+        azureOpenAIKey,
+        model,
+        temperature,
+        max_tokens,
+      }));
     }
   },
   updateEnvOpenAIKey: (envOpenAIKey) => {
@@ -91,9 +121,14 @@ const useStore = create<State & Action>((set) => ({
 }));
 
 const useOpenAI = (): UseOpenAIReturn => {
-  const { openAIKey, model, envOpenAIKey, temperature, max_tokens } = useStore(
-    (state) => state
-  );
+  const {
+    openAIKey,
+    azureOpenAIKey,
+    model,
+    envOpenAIKey,
+    temperature,
+    max_tokens,
+  } = useStore((state) => state);
 
   const update = useStore((state) => state.update);
 
@@ -101,13 +136,19 @@ const useOpenAI = (): UseOpenAIReturn => {
 
   React.useEffect(() => {
     const localOpenaiKey = localStorage.getItem("openaiKey") || "";
+    const localAzureOpenaiKey = localStorage.getItem("azureOpenAIKey") || "";
     const localTemperature = Number(localStorage.getItem("temperature") || "1");
     const localMaxToken = Number(localStorage.getItem("max_tokens") || "2000");
-    const localModel =
-      localStorage.getItem("language_model") || "gpt-3.5-turbo";
+    let localModel =
+      localStorage.getItem("language_model") || "openai-gpt-3.5-turbo";
+
+    if (!localModel.startsWith("openai") && !localModel.startsWith("azure")) {
+      localModel = "openai-gpt-3.5-turbo";
+    }
 
     update({
       openAIKey: localOpenaiKey,
+      azureOpenAIKey: localAzureOpenaiKey,
       temperature: isNaN(localTemperature) ? 1 : localTemperature,
       max_tokens: isNaN(localMaxToken) ? 2000 : localMaxToken,
       model: localModel,
@@ -118,6 +159,7 @@ const useOpenAI = (): UseOpenAIReturn => {
   return [
     {
       openAIKey,
+      azureOpenAIKey,
       model,
       envOpenAIKey,
       temperature,