+++ pages/api/openai.ts
@@ -5,7 +5,7 @@ export const config = {
   runtime: "edge",
 };
 
-const getProxyUrl = () => {
+const getEnvProxyUrl = () => {
   const API_PROXY = process.env.NEXT_PUBLIC_OPENAI_API_PROXY;
   if (!API_PROXY) return "";
   if (API_PROXY[API_PROXY.length - 1] === "/")
@@ -25,16 +25,21 @@ const handler = async (req: Request) => {
   const { proxyUrl, model, temperature, max_tokens, chat_list } =
     await req.json();
 
-  const findModel = modelOptions.find((m) => m.value === model);
+  const findModel = modelOptions.find((item) => {
+    return item.children.find((val) => val.value === model);
+  });
+
   if (!findModel) {
     return new Response("Error", { status: 500, statusText: "error" });
   }
 
-  const API_PROXY = getProxyUrl();
+  const ENV_API_PROXY = getEnvProxyUrl();
 
-  const proxy = proxyUrl || API_PROXY || "https://api.openai.com";
+  const proxy = proxyUrl || ENV_API_PROXY || "https://api.openai.com";
 
   const fetchURL = proxy + "/v1/chat/completions";
+
+  console.log(fetchURL, "fetchURL");
 
   const response = await fetch(fetchURL, {
     headers: {
@@ -44,7 +49,7 @@ const handler = async (req: Request) => {
     method: "POST",
     body: JSON.stringify({
       stream: true,
-      model,
+      model: model.split("-").slice(1).join("-"),
       temperature: isUndefined(temperature) ? 1 : temperature,
       max_tokens: isUndefined(max_tokens) ? 2000 : max_tokens,
       messages: [